v1: add new Disk partitioning support to API
Add the new Disk advanced partitioning to the API.  The full
specification of the Disk object and its sub-objects is copied from the
osbuild-composer cloud API.  For simplicity and to avoid needing to do
nested conversions, all the new objects have their x-go-type set to the
composer equivalents.  This makes it trivial to pass the pass the
options to osbuild-composer's API, while also allowing generating the
full API documentation.