internal/common: add helper for calculating sequence pagination
API endpoints returning collections / sequence must be paginated
according to the ConsoleDot API guidance. I noticed that some handlers
in the ib-crc v1 API that return paginated response do the calculation
of offsets incorrectly, but more importantly, each handler function
implements it in a slightly different way.

Let's add a helper for paginating sequence in a consistent and
unit-tested way.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>