v1: more flexible OpenSCAP customization comparison
v1.OpenSCAP is a union type, which means internally it's a raw message
([]byte) that can be unmarshalled into one of a number of types. If we
compare them directly, they wont match if the indentation of the json
doesn't match, which is now happening because of the change in
d069970faef34c59f3e4c62fa47ec1b4949f702f.

Marshalling the two instances makes sure we get rid of any internal json
formatting differences between the two sources.