distributions: generate package lists for rhel-9.6
895479b0066d5c3fe43a8f83c7e1bdf2624dfe8c mentioned that the package
lists are not needed because the frontend is now fully switched to use
the content service but we missed a piece: the edge service still uses
the package search provided by us.

This commit thus generates the package lists for 9.6 to fix the edge-mgmt
service. I ran the following command:

./tools/generate-package-lists \
  --distro distributions/rhel-9.6/rhel-9.6.json \
  --key /etc/pki/entitlement/ID-key.pem \
  --cert /etc/pki/entitlement/ID.pem

Note that generating the package list for RHEL 10.0 isn't necessary
because edge-mgmt doesn't supprot RHEL 10