v1/internal: Do not enable rhcd service for RHEL 10 (HMS-8756)
Because `rhcd` is no longer used in RHEL 10, enabling it causes RHEL 10
builds to fail.

The `rhcd` service is enabled for images that have an OpenSCAP profile.
This customization does not appear in the blueprint, rather it is
injected into the compose request.

This commit removes this behavior for RHEL 10.

We should also revisit enabling rhcd altogether in the future. It is
probably an antipattern. Consider:
1. What happens if a user explicitly masks or disables rhcd?
2. Should we even be doing this? We don't know why we need to do this.
3. If we do need to do this, shouldn't this knowledge live in
   osbuild/images?