v1: aap systemd services bug
When trying to enable the AAP first boot service we were causing
a panic and a 500 error in the service. The reason for this was
in cases where there was already a disabled or masked service,
the services struct was not nil but the enabled services were.
This was causing a nil pointer dereference and the panic.

This commit adds changes now to first check and, if necessary,
initialise the services struct. Then we can create a new struct
or safely append the first boot script to the list of
enabled services.