distributions: disable gpg checks for 9.7 and 10.1
RHEL 9.7 and 10.1 packages can be signed with a new PQC key.[1]

In an ideal world, we would just add this key to our distro defs,
and be done.

However, this key cannot be imported by rpmkeys from RHEL 9 that our
infra uses to build the images, so adding the key will break the builds.
This effectively means that we cannot currently add the new key.
The only remaining option is to disable the gpg key checking before
we are able to figure out a proper fix.

Is this still secure? I believe so because we trust cdn.redhat.com,
and any MITM attacks should be mitigated by the TLS layer. Yes, this
removes one layer of protection, but it's only a temporary measure.

[1]: https://access.redhat.com/security/team/key