internal/v1: Improve 'quota exceeded' error message
This commit adds a new QuotaResult struct to provide more context to
users in the error message whenever their quota is exceeded. Since we
don't surface the actual quota or the usage against the quota anywhere
in the user interface, this feels like helpful additional context for
customers.
In order to provide this context, the internal/common/quota.go now
returns a struct (Ok, Limit, Used, SlidingWindow) instead of a simple
boolean.

Currently, the error message for when customers exceed their build quota
simply reads:
  "Quota exceeded for user"

With the proposed changes, the error message will read:
  "Quota of 100 builds per 14 days exceeded for user"